+++ recipes/jna.rb
@@ -1,8 +1,8 @@
 #
-## Cookbook Name:: dse-nativex
+## Cookbook Name:: dse-simplyadrian
 ## Recipe:: jna.rb
 ##
-## Copyright 2014, NativeX
+## Copyright 2014, simplyadrian
 ##
 ## All rights reserved - Do Not Redistribute
 ##